make the tests to work with tls
