+++ danube-broker/tests/reliable_dispatch_shared.rs
@@ -3,27 +3,42 @@ extern crate futures_util;
 
 use anyhow::Result;
 use danube_client::{
-    ConfigReliableOptions, ConfigRetentionPolicy, Consumer, DanubeClient, Producer, SubType,
+    ConfigReliableOptions, ConfigRetentionPolicy, ConnectionOptions, Consumer, DanubeClient,
+    Producer, SubType,
 };
 use rustls::crypto;
 use std::fs;
 use std::sync::Arc;
+use tokio::sync::OnceCell;
 use tokio::time::{sleep, timeout, Duration};
+use tonic::transport::{Certificate, ClientTlsConfig};
+
+static CRYPTO_PROVIDER: OnceCell<()> = OnceCell::const_new();
 
 struct TestSetup {
     client: Arc<DanubeClient>,
 }
 
 async fn setup() -> Result<TestSetup> {
-    // Install the default CryptoProvider
-    let crypto_provider = crypto::ring::default_provider();
-    crypto_provider
-        .install_default()
-        .expect("Failed to install default CryptoProvider");
+    CRYPTO_PROVIDER
+        .get_or_init(|| async {
+            let crypto_provider = crypto::ring::default_provider();
+            crypto_provider
+                .install_default()
+                .expect("Failed to install default CryptoProvider");
+        })
+        .await;
+
+    let tls_config = ClientTlsConfig::new().ca_certificate(Certificate::from_pem(
+        std::fs::read("../cert/ca-cert.pem").unwrap(),
+    ));
+
+    let connection_options = ConnectionOptions::new().tls_config(tls_config);
 
     let client = Arc::new(
         DanubeClient::builder()
-            .service_url("http://127.0.0.1:6650")
+            .service_url("https://127.0.0.1:6650")
+            .with_connection_options(connection_options)
             .build()
             .await?,
     );
@@ -36,12 +51,8 @@ async fn setup_reliable_producer(
     topic: &str,
     producer_name: &str,
 ) -> Result<Producer> {
-    // Create a reliable dispatch options with 1 hour retention and 5MB segment size
-    let reliable_options = ConfigReliableOptions::new(
-        5, // segment size in MB
-        ConfigRetentionPolicy::RetainUntilExpire,
-        3600, // 1 hour
-    );
+    let reliable_options =
+        ConfigReliableOptions::new(5, ConfigRetentionPolicy::RetainUntilExpire, 3600);
 
     let mut producer = client
         .new_producer()
@@ -51,7 +62,6 @@ async fn setup_reliable_producer(
         .build();
 
     producer.create().await?;
-
     Ok(producer)
 }
 
@@ -74,66 +84,9 @@ async fn setup_reliable_consumer(
 }
 
 #[tokio::test]
-async fn test_reliable_exclusive_delivery() -> Result<()> {
-    let setup = setup().await?;
-    let topic = "/default/topic_test_reliable_exclusive";
-
-    // Read the blob file
-    let blob_data = fs::read("./tests/test.blob")?;
-
-    let producer = setup_reliable_producer(
-        setup.client.clone(),
-        topic,
-        "test_producer_reliable_exclusive",
-    )
-    .await?;
-
-    let mut consumer = setup_reliable_consumer(
-        setup.client.clone(),
-        topic,
-        "test_consumer_reliable_exclusive",
-        SubType::Exclusive,
-    )
-    .await?;
-
-    let mut message_stream = consumer.receive().await?;
-
-    sleep(Duration::from_millis(500)).await;
-
-    // Send blob 20 times
-    for i in 0..20 {
-        let blob_clone = blob_data.clone();
-        let message_id = producer.send(blob_clone, None).await?;
-        println!("Blob message {} sent with id: {}", i, message_id);
-    }
-
-    let receive_future = async {
-        for i in 0..20 {
-            if let Some(stream_message) = message_stream.recv().await {
-                //let payload = String::from_utf8(stream_message.payload.clone()).unwrap();
-                assert_eq!(stream_message.payload, blob_data);
-                println!("Received blob message #{}", i);
-                consumer.ack(&stream_message).await?;
-            }
-        }
-        Ok::<(), anyhow::Error>(())
-    };
-
-    let result = timeout(Duration::from_secs(10), receive_future).await;
-    assert!(
-        result.is_ok(),
-        "Test timed out while waiting for the message"
-    );
-
-    Ok(())
-}
-
-#[tokio::test]
 async fn test_reliable_shared_delivery() -> Result<()> {
     let setup = setup().await?;
     let topic = "/default/topic_test_reliable_shared";
-
-    // Read the blob file
     let blob_data = fs::read("./tests/test.blob")?;
 
     let producer =
@@ -152,7 +105,6 @@ async fn test_reliable_shared_delivery() -> Result<()> {
 
     sleep(Duration::from_millis(500)).await;
 
-    // Send blob 20 times
     for i in 0..20 {
         let blob_clone = blob_data.clone();
         let message_id = producer.send(blob_clone, None).await?;
@@ -162,7 +114,6 @@ async fn test_reliable_shared_delivery() -> Result<()> {
     let receive_future = async {
         for i in 0..20 {
             if let Some(stream_message) = message_stream.recv().await {
-                //let payload = String::from_utf8(stream_message.payload.clone()).unwrap();
                 assert_eq!(stream_message.payload, blob_data);
                 println!("Received blob message #{}", i);
                 consumer.ack(&stream_message).await?;
